requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
tqdm>=4.66.0
colorama>=0.4.6
facenet-pytorch>=2.5.3
